Reduced CPU max grid size to match Cuda accelerators.

+++ Src/ILGPU/Runtime/CPU/CPUDevice.cs
@@ -233,7 +233,7 @@ namespace ILGPU.Runtime.CPU
                 MaxNumThreadsPerGroup);
 
             MemorySize = long.MaxValue;
-            MaxGridSize = new Index3(int.MaxValue, int.MaxValue, int.MaxValue);
+            MaxGridSize = new Index3(int.MaxValue, ushort.MaxValue, ushort.MaxValue);
             MaxSharedMemoryPerGroup = int.MaxValue;
             MaxConstantMemory = int.MaxValue;
             NumThreads = MaxNumThreads * numMultiprocessors;
